refactor(dftu): split contributeHR and extract kernel functions
dftu_nao_op.cpp was 720 lines with contributeHR CC=20.

1. Split contributeHR into three private member functions:
   - compute_occ_from_dmr: BRANCH 1 (DMR-based occ calculation)
   - load_occ_from_file: BRANCH 2 (pre-read occ_mat loading)
   - accumulate_HR_for_iat0: Step 5 (HR pair accumulation)
   contributeHR now reads as a thin (T,I,l,n) walk calling the helpers;
   CC drops 20 -> ~4.

2. Extract cal_HR_IJR, cal_occ, transfer_pot_onsite (both generic and
   noncollinear specialization), and cal_pot_onsite into a new
   dftu_nao_op_kernel.cpp translation unit. op.cpp drops from 720 to
   ~440 lines, under the 500-line limit. The new file carries its own
   explicit template instantiations matching op.cpp. The test CMakeLists
   is updated to compile the new source alongside the existing ones.

Scores: op.cpp 72->95, kernel 97. Build clean.
